Update Files doc
The Finder integration is not on by default any more. So mention the extra "install" step to avoid confusion. Also, drag/drop works in the GUI app (i.e. without finder integration as well), but I'm not sure how to clarify that without changing the subheader "Drag and Drop" which I also like.